requirements.txt
emoji==1.7.0
numpy==1.19.5
pandas==1.1.5
pymongo==3.6.1
python-dotenv==0.20.0
scikit_learn==1.0.2
scipy==1.5.4
textblob==0.17.1
tqdm==4.64.0
tweepy==3.9.0
